pyproject.toml
[project]
name = "langgraph-learning"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.3.27",
    "langchain-google-genai>=2.1.9",
    "langgraph>=0.6.3",
    "langsmith>=0.4.11",
    "notebook>=7.4.5",
    "python-dotenv>=1.1.1",
]
